// ichub-frontend/src/features/industry-core-kit/part-discovery/components/submodel-addons/battery-pass-technical-data/types.ts
export type BatteryCategory = 'lmt' | 'ev' | 'industrial' | 'stationary';

export interface GeneralInformation {
  ManufacturerName: string;
  CompanyLogo?: string;
  ManufacturerProductDesignation: MultiLangEntry[];
  ManufacturerArticleNumber: string;
  ManufacturerOrderCode: string;
  ManufacturerIdentifier: string;
  WarrantyPeriod: string;
  BatteryCategory: BatteryCategory;
  BatteryMass: number;
}

export interface PowerCapabilityAt {
  atSoC: number;
  powerCapabilityAt: number;
}

export interface CapacityEnergyVoltage {
  NominalVoltage: number;
  MinVoltage: number;
  MaxVoltage: number;
  RatedCapacity: number;
  CapacityFade?: number;
  CertifiedUsableBatteryEnergy?: number;
}

export interface RoundTripEnergyEfficiency {
  InitialRoundTripEnergyEfficiency: number;
  RoundTripEnergyEfficiencyAt50PercentOfCycleLife: number;
  EnergyRoundTripEfficiencyFade?: number;
  InitialSelfDischargingRate?: number;
}

export interface Resistance {
  InitialInternalResistanceAtBatteryCellLevel: number;
  InitialInternalResistanceAtBatteryPackLevel: number;
  InitialInternalResistanceAtBatteryModuleLevel?: number;
  InternalResistanceIncreaseAtBatteryCellLevel?: number;
  InternalResistanceIncreaseAtBatteryPackLevel: number;
  InternalResistanceIncreaseAtBatteryModuleLevel?: number;
}

export interface PowerCapability {
  MaximumPermittedBatteryPower: number;
  PowerFade: number;
  RatioNominalBatteryPowerAndBatteryEnergy: number;
  OriginalPowerCapability: PowerCapabilityAt[];
}

export interface Temperature {
  TemperatureRangeIdleState_LowerBoundary: number;
  TemperatureRangeIdleState_UpperBoundary: number;
}

export interface Lifetime {
  ExpectedLifetimeInCalendarYears: number;
  ExpectedNumberOfCycles: number;
  CapacityThresholdExhaustion: number;
  CycleLifeReferenceTest: string[];
  CrateOfRelevantCycleLifeTest: number;
}

export interface TechnicalAttributes {
  CapacityEnergyVoltage: CapacityEnergyVoltage;
  RoundTripEnergyEfficiency: RoundTripEnergyEfficiency;
  Resistance: Resistance;
  PowerCapability: PowerCapability;
  Temperature: Temperature;
  Lifetime: Lifetime;
}

export interface TechnicalData {
  GeneralInformation: GeneralInformation;
  TechnicalPropertyAreas: TechnicalAttributes;
}